Add (somewhat incomplete) support for SERVER_TRANSPORT_OPTIONS.
To be sepcific, this still needs to handle unescaping values, and needs unit
tests.  Not sure if it's worth bothering scrubbing the SERVER_TRANSPORT_OPTIONS
when freeing either since the values will be stuck in envp anyway since there is
no supported method of scrubbing that buffer, so I might revert where I do that,
though I will keep the utility function that's part of the library.